adiciona NomeEAbreviacaoUnidadeMedidaSerializer
