+++ sme_terceirizadas/pre_recebimento/api/serializers/serializers.py
@@ -217,14 +217,11 @@ class UnidadeMedidaSerialzer(serializers.ModelSerializer):
     class Meta:
         model = UnidadeMedida
         fields = ('uuid', 'nome', 'abreviacao', 'criado_em')
-        read_only_fields = ('uuid', 'criado_em')
+        read_only_fields = ('uuid', 'nome', 'abreviacao', 'criado_em')
 
-    def validate_nome(self, value):
-        if not value.isupper():
-            raise serializers.ValidationError('O campo deve conter apenas letras maiúsculas.')
-        return value
 
-    def validate_abreviacao(self, value):
-        if not value.islower():
-            raise serializers.ValidationError('O campo deve conter apenas letras minúsculas.')
-        return value
+class NomeEAbreviacaoUnidadeMedidaSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = UnidadeMedida
+        fields = ('nome', 'abreviacao')
+        read_only_fields = ('nome', 'abreviacao')
